refactor: extract engineTypes.ts to break circular dependency
Move JsonCompletionEngine, EnginePolicy, and EngineUnavailableError into
engineTypes.ts. Both engines.ts and localMlEngine.ts import from it,
breaking the engines ↔ localMlEngine cycle caught by madge CI check.

engines.ts re-exports all types for backward compatibility — no consumer
import changes needed outside ai-engine package.
